novelist/pyproject.toml
[project]
name = "novelist"
version = "0.1.0"
description = "Scientific Hypothesis Synthesizer - A research soul that iteratively generates, debates, verifies, and refines scientific hypotheses"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
    "respx>=0.21.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
]

[project.scripts]
novelist = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
